requirements.txt
arabic-reshaper
asgiref
asn1crypto
atomicwrites
attrs
Babel
certifi
cffi
chardet
charset-normalizer
click
colorama
colorlog
coreapi
coreschema
coverage
cryptography
cssselect2
Django
django-cors-headers
django-extensions
django-filter
django-sequences
django-simple-history
djangorestframework
docopt
drf-writable-nested
drf-yasg
drf-yasg2
et-xmlfile
execnet
factory-boy
Faker
gunicorn
mock
html5lib
idna
inflection
iniconfig
itypes
Jinja2
lxml
MarkupSafe
num2words
numpy
numpy-financial
openpyxl
oscrypto
packaging
pandas
pillow
pluggy
psycopg2
psycopg2-binary
django-heroku
py
pycparser
pyHanko
pyhanko-certvalidator
PyJWT
PyMuPDF
PyMuPDFb
pyparsing
pypdf
pypng
pytest
pytest-cov
pytest-django
pytest-forked
pytest-xdist
python-bidi
python-dateutil
pytz
PyYAML
reportlab
requests
requests-mock
responses
ruamel.yaml
ruamel.yaml.clib
setuptools
six
sqlparse
svglib
tinycss2
tomli
typing_extensions
tzdata
tzlocal
uritemplate
uritools
urllib3
setuptools
webencodings
xhtml2pdf
sib_api_v3_sdk
python-dotenv
whitenoise

